@@ -23,20 +23,23 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  font_awesome_flutter: ^8.4.0
-  path_provider: ^2.0.1
+  font_awesome_flutter: ^9.0.0
+  path_provider:
   toast: ^0.1.5
-  http: ^0.12.2
-  url_launcher: ^5.2.5
-  intl: ^0.16.1
+  http:
+  url_launcher:
+  intl: ^0.17.0
   date_picker_timeline: ^1.2.1
-  flutter_staggered_grid_view: ^0.3.3
+  flutter_staggered_grid_view: ^0.4.0
   flutter_sparkline: ^0.1.0
   cupertino_icons: ^1.0.2
   chips_choice: ^2.0.1
-  firebase_core: ^1.0.3
+  firebase_core:
   cloud_firestore:
   provider:
+  passcode_screen: ^2.0.0
+  uuid: ^2.2.2
+  expansion_card:
 
 dev_dependencies:
   flutter_test:
