@@ -1,8 +1,11 @@
+import 'dart:convert';
 import 'dart:math';
 
 import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:delivery_santanna/models/cart.dart';
 import 'package:delivery_santanna/models/product.dart';
 import 'package:delivery_santanna/models/order_store.dart';
+import 'package:delivery_santanna/models/exception_event.dart';
 
 import 'dao.dart';
 
@@ -12,6 +15,7 @@ class CRUDModel{
 
   Dao _dao;
   List<Product> products;
+  List<OrderStore> customerOrders;
 
   CRUDModel(this.collection){
     _dao = Dao(this.collection);
@@ -25,6 +29,26 @@ class CRUDModel{
         .toList();
 
     return products;
+  }
+
+  Future<List<OrderStore>> fetchCustomersOrder() async {
+
+    try{
+      var result = await _dao.getOrdersStoreCollection();
+
+      customerOrders = result.docs
+          .map((doc) => OrderStore.fromMap(
+          doc.data(),
+          doc.id,
+          buildListCart(doc.data()['cartItemsList'])))
+          .toList();
+
+      print(customerOrders);
+
+      return customerOrders.toList();
+    }catch(e){
+      throw Exception(e);
+    }
   }
 
   Stream<QuerySnapshot> fetchProductsAsStream() {
@@ -49,25 +73,82 @@ class CRUDModel{
   Future addProduct(Product data) async{
 
     await _dao.addDocument(data.toJson());
-
     return ;
 
   }
 
-  Future addOrder(
-     String name,
-     String total,
-     String time,
-     String order) async {
+  Future addException(
+      String name,
+      String exception,
+      String time) async {
 
-    OrderStore orderStore = OrderStore(
+    ExceptionEvent exceptionEv = ExceptionEvent(
         Random.secure().nextInt(40000000).toString(),
         name,
-        order,
+        exception,
+        time);
+
+    await _dao.addDocument(exceptionEv.toJson());
+    return ;
+  }
+
+  Future addOrder(
+      String uniqueId,
+      String name,
+      String total,
+      String time,
+      List<dynamic> cartItems,
+      bool confirmed,
+      String typeOrder,
+      String datePickupDelivery,
+      String hourPickupDelivery,
+      String address,
+      String city) async {
+
+    OrderStore orderStore = OrderStore(
+        '',
+        uniqueId,
+        name,
+        cartItems,
         time,
-      total);
+        total,
+        confirmed,
+        typeOrder,
+        datePickupDelivery,
+        hourPickupDelivery,
+        city,
+        address);
 
     await _dao.addDocument(orderStore.toJson());
     return ;
+  }
+
+  List buildListCart(List element) {
+
+    try{
+      List<Cart> listCart = <Cart>[];
+      element.forEach((currentItem) {
+
+        currentItem = currentItem.toString().replaceAll('{', '{"');
+        currentItem = currentItem.toString().replaceAll('}', '"}');
+        currentItem = currentItem.toString().replaceAll(',', '","');
+        currentItem = currentItem.toString().replaceAll(':', '":"');
+        currentItem = currentItem.toString().replaceAll(' product', 'product');
+        currentItem = currentItem.toString().replaceAll(' numberOfItem', 'numberOfItem');
+        currentItem = currentItem.toString().replaceAll(' changes', 'changes');
+        Map valueMap = json.decode(currentItem);
+        listCart.add(Cart(
+            product : Product('', valueMap['product'], '', ["-"], ["-"], 0.0, 0, ["-"], '', 'true'),
+            numberOfItem: int.parse(valueMap['numberOfItem'].toString().replaceAll(" ", "")),
+            changes: null
+        ));
+      });
+
+      return listCart;
+    }catch(e){
+      print('Exception: ' + e.toString());
+      throw Exception(e);
+    }
+
   }
 }
