@@ -1,36 +1,96 @@
+import 'package:delivery_santanna/models/cart.dart';
+
 class OrderStore {
 
+  final String docId;
   final String id;
   final String name;
-  final String order;
+  final List<dynamic> cartItemsList;
   final String date;
   final String total;
-
+  bool confirmed;
+  final String typeOrder;
+  final String datePickupDelivery;
+  final String hourPickupDelivery;
+  final String city;
+  final String address;
 
   OrderStore(
+      this.docId,
       this.id,
       this.name,
-      this.order,
+      this.cartItemsList,
       this.date,
-      this.total
+      this.total,
+      this.confirmed,
+      this.typeOrder,
+      this.datePickupDelivery,
+      this.hourPickupDelivery,
+      this.city,
+      this.address,
       );
 
 
   toJson(){
-
     return {
+      'docId' : docId,
       'id' : id,
       'name' : name,
-      'message': order,
+      'cartItemsList': serializeList(cartItemsList),
       'date': date,
-      'total' : total + ' €'
+      'total' : total + ' €',
+      'confirmed' : confirmed,
+      'typeOrder' : typeOrder,
+      'datePickupDelivery' : datePickupDelivery,
+      'hourPickupDelivery' : hourPickupDelivery,
+      'city' : city,
+      'address' : address
     };
   }
 
+  factory OrderStore.fromMap(
+      Map snapshot,
+      String docId,
+      List<Cart> cartList
+      ){
+    return OrderStore(
+      docId,
+      snapshot['id'] as String,
+      snapshot['name'] as String,
+      cartList,
+      snapshot['date'] as String,
+      snapshot['total'] as String,
+      snapshot['confirmed'] as bool,
+      snapshot['typeOrder'] as String,
+      snapshot['datePickupDelivery'] as String,
+      snapshot['hourPickupDelivery'] as String,
+      snapshot['city'] as String,
+      snapshot['address'] as String,
+    );
+  }
+
+  serializeList(List<Cart> cartItemsList) {
+    List<String> orderDeserialized = <String>[];
+
+    cartItemsList.forEach((element) {
+      print('AUAUSADUASDUASUDASDU');
+      print(element.toJson());
+      print('AUAUSADUASDUASUDASDU');
+      orderDeserialized.add(element.toJson().toString());
+    });
+    return orderDeserialized;
+  }
 
   @override
   String toString() {
-    return this.order;
+    return 'Name: ' + name + ' - Hour PickupDelivery: '
+        + hourPickupDelivery + ' '
+        '- Cart: ' + cartItemsList.toString();
   }
+
+  void buildCartListFromListOfString(String cartList) {
+    print(cartList);
+  }
+
 
 }
